Use cheap API call to verify access token
It seems the ```client.fetch("1", {})``` call was used only for checking for any `Unauthorized - Session not found` errors. Since the `*` query returns all documents the token has access to, this operation could get expensive for some users. Changing this query to return something lighter, like `1` will achieve the same, but in a much lighter way.